+++ django_spire/core/management/commands/spire_startapp_pkg/exceptions.py
@@ -1,0 +1,5 @@
+from django.core.management import CommandError
+
+
+class AppExistsError(CommandError):
+    pass